create header container, code refactoring
